% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_stminsights}
\alias{run_stminsights}
\title{This is a helper function to launch the stminsights app}
\usage{
run_stminsights()
}
\description{
\code{run_stminsights} launches the app to analyze Structural Topic models.
It requires uploading an R image containing an arbitrary number of stm models and effect estimates.
In addition, the image should contain the \code{out} object which was used to fit the models.
See the description for additional information.
}
\details{
To be written (sorry =D).
}
