% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_diag.R
\name{stm_diag}
\alias{stm_diag}
\title{computes stm model diagnostics}
\usage{
stm_diag(models, outobj)
}
\arguments{
\item{models}{A list of stm models.}

\item{outobj}{The \code{out} object containing documents for all stm models.}
}
\value{
Returns model diagnostics in a tidy data frame.
}
\description{
\code{ stm_diag()} is a helper function to compute average and median
\code{\link[stm]{semanticCoherence}} and \code{\link[stm]{exclusivity}} for
a number of  \code{\link[stm]{stm}} models. The function does not work for
models with content covariates.
}
\examples{
\dontrun{
library(stm)
library(dplyr)
library(ggplot2)

# out object
out <- list(documents = poliblog5k.docs,
            vocab = poliblog5k.voc,
            meta = poliblog5k.meta)

# one or several stm models
poli10 <- stm(documents = out$documents,
            vocab = out$vocab,
            data = out$meta,
            prevalence = ~ rating * s(day),
            K = 10)
poli20 <- stm(documents = out$documents,
            vocab = out$vocab,
            data = out$meta,
            prevalence = ~ rating * s(day),
            K = 20)
poli30 <- stm(documents = out$documents,
            vocab = out$vocab,
            data = out$meta,
            prevalence = ~ rating * s(day),
            K = 30)

# get diagnostics
diag <- stm_diag(models = list(
                 model_10 = poli10,
                 model_20 = poli20,
                 model_30 = poli30),
                 outobj = out)

# plot diagnostics
diag \%>\%
ggplot(aes(x = coherence, y = exclusivity, color = statistic))  +
  geom_text(aes(label = name), nudge_x = 0.8) + geom_point() +
  labs(x = 'Semantic Coherence', y = 'Exclusivity') + theme_light()
}
}
