% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm_effects.R
\name{stm_effects}
\alias{stm_effects}
\title{extract stm effect estimates}
\usage{
stm_effects(estimates, variable, type, ci = 0.95, moderator = NULL,
  modval = NULL)
}
\arguments{
\item{estimates}{The object containing estimates calculated with
\code{\link[stm]{estimateEffect}}.}

\item{variable}{The variable for which estimates should be extracted.}

\item{type}{The estimate type. Can either be \code{'pointestimate'}
or  \code{'continuous'}.}

\item{ci}{The confidence interval for uncertainty estimates.
Defaults to  \code{0.95}.}

\item{moderator}{The moderator variable in case you want to include
an interaction effect.}

\item{modval}{The value of the moderator variable for an interaction effect.
See examples for combining data for multiple values.}
}
\value{
Returns effect estimates in a tidy data frame.
}
\description{
\code{ stm_effects()} is a helper function to store effect estimates from
stm in a data frame.
}
\examples{
\dontrun{

library(stm)
library(dplyr)
library(ggplot2)

# store effects
prep <- estimateEffect(1:3 ~ treatment, gadarianFit, gadarian)

effects <- stm_effects(estimates = prep,
                      variable = 'treatment',
                      type = 'pointestimate')

# plot effects
effects \%>\% filter(topic == 3) \%>\%
ggplot(aes(x = value, y = proportion)) +
 geom_errorbar(aes(ymin = lower, ymax = upper), width = 0.1, size = 1) +
 coord_flip() + theme_light() + labs(x = 'Treatment', y = 'Topic Proportion')

# combine estimates for interaction effects
prep_int <- estimateEffect(1:3 ~ treatment*s(pid_rep), gadarianFit, gadarian)

effects_int <- stm_effects(estimates = prep_int,
                          variable = 'pid_rep',
                          type = 'continuous',
                          moderator = 'treatment',
                          modval = 1) \%>\%
 bind_rows(
   stm_effects(estimates = prep_int,
               variable = 'pid_rep',
               type = 'continuous',
               moderator = 'treatment',
               modval = 0)
 )
}

}
