% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_stminsights}
\alias{run_stminsights}
\title{launch the stminsights shiny app}
\usage{
run_stminsights(use_browser = TRUE)
}
\arguments{
\item{use_browser}{Choose whether you want to launch the shiny app in your browser.
Defaults to \code{TRUE}.}
}
\description{
\code{run_stminsights} launches the app to analyze Structural Topic models.
It requires an .RData file with stm objects as illustrated in the example below.
}
\examples{

 \dontrun{

library(stm)
library(quanteda)

# prepare data
data <- corpus(gadarian, text_field = 'open.ended.response')
docvars(data)$text <- texts(data)
data <- dfm(data, stem = TRUE, remove = stopwords('english'),
            remove_punct = TRUE) \%>\% dfm_trim(min_count = 2)
out <- convert(data, to = 'stm')

# fit models and effect estimates
gadarian_3 <- stm(documents = out$documents,
                  vocab = out$vocab,
                  data = out$meta,
                  prevalence = ~ treatment + s(pid_rep),
                  K = 3,
                  max.em.its = 1, # reduce computation time for example
                  verbose = FALSE)

prep_3 <- estimateEffect(1:3 ~ treatment + s(pid_rep), gadarian_3,
                         meta = out$meta)

gadarian_5 <- stm(documents = out$documents,
                  vocab = out$vocab,
                  data = out$meta,
                  prevalence = ~ treatment + s(pid_rep),
                  K = 5,
                  max.em.its = 1, # reduce computation time for example
                  verbose = FALSE)

prep_5 <- estimateEffect(1:5 ~ treatment + s(pid_rep), gadarian_5,
                         meta = out$meta)

# save objects in .RData file
save.image(paste0(tempdir(), '/stm_gadarian.RData'))

# launch the app
if(interactive()){
  run_stminsights()
}

}

}
